Tidy Up MyPatients
deleted previous router for MyPastPatients, created components for UpcomingPatient and PastPatient to be accessed directly from MyPatients instead
